Made the ViewDefinition test standalone.
